refactor: update package name to match subsystem name
